IAT-3078: Refactored most of validation into common and out of the validation app.
