+++ src/test/java/org/opentestsystem/ap/ivs/AppTestConfig.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.opentestsystem.ap.ivs.config;
+package org.opentestsystem.ap.ivs;
 
 import org.gitlab4j.api.GitLabApi;
 import org.gitlab4j.api.models.Namespace;
