+++ src/test/java/org/opentestsystem/ap/ivs/IVSApplicationIT.java
@@ -10,7 +10,7 @@
  * specific language governing permissions and limitations under the license.
  */
 
-package org.opentestsystem.ap.ivs.it;
+package org.opentestsystem.ap.ivs;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
