+++ src/test/java/org/opentestsystem/ap/ivs/ValidationApiTest.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.ap.ivs.rest.v1;
+package org.opentestsystem.ap.ivs;
 
 
 import org.junit.Before;
@@ -8,7 +8,6 @@ import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.opentestsystem.ap.common.model.validation.ValidationResults;
 import org.opentestsystem.ap.common.validation.ValidationService;
-import org.opentestsystem.ap.ivs.ValidationApi;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
