+++ src/test/java/org/opentestsystem/ap/ivs/ValidationUtilityTest.java
@@ -10,7 +10,7 @@
  * specific language governing permissions and limitations under the license.
  */
 
-package org.opentestsystem.ap.ivs.service;
+package org.opentestsystem.ap.ivs;
 
 import org.apache.commons.io.FileUtils;
 import org.junit.After;
@@ -75,7 +75,7 @@ public class ValidationUtilityTest {
 
         clonedStimPath = baseDir.resolve(itemRootFolderStr + "/" + LEGACY_STIM_ID);
 
-        utility = new ValidationUtility(testHelper.getItemBankProperties(), testHelper.getIvsProperties());
+        utility = new ValidationUtility(testHelper.getIvsProperties());
 
         service = new ValidationService(
             validationManager,
